+++ builder_pkgs/mockito/test/end2end/foo.dart
@@ -1,0 +1,5 @@
+class Foo {
+  String m(Bar bar) => 'result';
+}
+
+class Bar {}